Storytelling/requirements.txt
fastapi
uvicorn[standard]
requests
firebase-admin
python-dotenv
